PRSD-NONE: review mark-ups
