+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/local/api/mockOneLoginResponses/token.json
@@ -1,0 +1,6 @@
+{
+  "access_token": "SlAV32hkKG",
+  "token_type": "Bearer",
+  "expires_in": 180,
+  "id_token": "idToken"
+}
