+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/local/api/mockOneLoginResponses/userInfo.json
@@ -1,0 +1,7 @@
+{
+  "sub": "userId",
+  "email": "userEmail",
+  "email_verified": true,
+  "phone_number": "userNumber",
+  "phone_number_verified": true
+}
